+++ src/main/java/com/sparta/hhplusconcert/queue/domain/Status.java
@@ -1,4 +1,4 @@
-package com.sparta.hhplusconcert.domain.common;
+package com.sparta.hhplusconcert.queue.domain;
 
 import lombok.Getter;
