+++ src/main/java/com/sparta/hhplusconcert/queue/usecase/GetRemainingQueueService.java
@@ -1,6 +1,6 @@
 package com.sparta.hhplusconcert.queue.usecase;
 
-import com.sparta.hhplusconcert.domain.common.Status;
+import com.sparta.hhplusconcert.queue.domain.Status;
 import com.sparta.hhplusconcert.queue.domain.ExpiredTokenException;
 import com.sparta.hhplusconcert.queue.domain.InvalidTokenException;
 import com.sparta.hhplusconcert.queue.domain.entity.QueueTokenEntity;
